feat: add mongodb@4.x support
BREAKING CHANGE: it is hard to support both versions (3.x and 4.x) of
native driver on the same branch because 4.x has native types
definitions and they diverge from the ones provided by @types/mongodb.

This means that this version might still be compatible with previous
versions of mongodb native driver (3.6.x and 3.7.x), but we aren't
testing them due to typescript mismatch.

Please open an issue and we will try to support your use case.